# erp-domain/openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:51904
  description: Generated server url
paths:
...
  /api/v1/orchestrator/payroll/run:
    post:
      tags:
      - orchestrator-controller
      operationId: runPayroll
      parameters:
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/orders/{orderId}/fulfillment:
    post:
      tags:
      - orchestrator-controller
      operationId: fulfillOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderFulfillmentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/orders/{orderId}/approve:
    post:
      tags:
      - orchestrator-controller
      operationId: approveOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/factory/dispatch/{batchId}:
    post:
      tags:
      - orchestrator-controller
      operationId: dispatch
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/dispatch:
    post:
      tags:
      - orchestrator-controller
      operationId: dispatchOrder
      parameters:
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/dispatch/{orderId}:
    post:
      tags:
      - orchestrator-controller
      operationId: dispatchOrderAlias
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
  /api/v1/orchestrator/traces/{traceId}:
    get:
      tags:
      - orchestrator-controller
      operationId: trace
      parameters:
      - name: traceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
components:
  schemas:
...
    PayrollRunRequest:
      required:
      - creditAccountId
      - debitAccountId
      - initiatedBy
      - payrollDate
      - postingAmount
      type: object
      properties:
        payrollDate:
          type: string
          format: date
        initiatedBy:
          type: string
        debitAccountId:
          type: integer
          format: int64
        creditAccountId:
          type: integer
          format: int64
        postingAmount:
          type: number
    OrderFulfillmentRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
    ApproveOrderRequest:
      required:
      - approvedBy
      - orderId
      - totalAmount
      type: object
      properties:
        orderId:
          type: string
        approvedBy:
          type: string
        totalAmount:
          type: number
    DispatchRequest:
      required:
      - batchId
      - postingAmount
      - requestedBy
      type: object
      properties:
        batchId:
          type: string
        requestedBy:
          type: string
        postingAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
...